requirements.txt
Flask
h5py
numpy
Pillow
tensorflow
timm
gradio_client